+++ api/src/models/fujifilm/str/mod.rs
